experiment with mashumaro
