+++ core/dbt/contracts/relation.py
@@ -5,7 +5,7 @@ from typing import (
 )
 from typing_extensions import Protocol
 
-from hologram import JsonSchemaMixin
+from dbt.contracts.jsonschema import JsonSchemaMixin
 from hologram.helpers import StrEnum
 
 from dbt import deprecations
@@ -13,8 +13,6 @@ from dbt.contracts.util import Replaceable
 from dbt.exceptions import CompilationException
 from dbt.utils import deep_merge
 
-
-from mashumaro import DataClassJSONMixin
 
 class RelationType(StrEnum):
     Table = 'table'
@@ -33,40 +31,8 @@ class ComponentName(StrEnum):
 class HasQuoting(Protocol):
     quoting: Dict[str, bool]
 
-"""
-This is a throwaway shim to match the JsonSchemaMixin interface
-that downstream consumers (ie. PostgresRelation) is expecting. 
-I imagine that we would try to remove code that depends on this type
-reflection if we pursue an approach like the one shown here
-"""
-class FastDataClassJSONMixing(DataClassJSONMixin):
-    @classmethod
-    def field_mapping(cls) -> Dict[str, str]:
-        """Defines the mapping of python field names to JSON field names.
 
-        The main use-case is to allow JSON field names which are Python keywords
-        """
-        return {}
-
-    @classmethod
-    def _get_fields(cls) -> List[Tuple[Field, str]]:
-        mapped_fields = []
-        type_hints = get_type_hints(cls)
-
-        for f in fields(cls):
-            # Skip internal fields
-            if f.name.startswith("_"):
-                continue
-
-            # Note fields() doesn't resolve forward refs
-            f.type = type_hints[f.name]
-
-            mapped_fields.append((f, cls.field_mapping().get(f.name, f.name)))
-
-        return mapped_fields  # type: ignore
-
-
-class FakeAPIObject(FastDataClassJSONMixing, Replaceable, Mapping):
+class FakeAPIObject(JsonSchemaMixin, Replaceable, Mapping):
     # override the mapping truthiness, len is always >1
     def __bool__(self):
         return True
@@ -92,8 +58,12 @@ class FakeAPIObject(FastDataClassJSONMixing, Replaceable, Mapping):
         return self.from_dict(value)
 
 
-T = TypeVar('T')
-
+"""
+mashumaro does not support generic types, so I just duplicated
+the generic type class that was here previously to get things
+working. We'd probably want to do this differently
+in the future
+"""
 
 @dataclass
 class Policy(FakeAPIObject):
